feat: event hub to db

+++ monicar-event-hub/src/main/java/org/eventhub/infrastructure/messaging/command/CycleInfoCommand.java
@@ -13,8 +13,8 @@ import lombok.Builder;
 public record CycleInfoCommand(
 	@JsonFormat(pattern = "yyyyMMddHHmmss") LocalDateTime intervalAt,
 	GpsStatus gcd,
-	Long lat,
-	Long lng,
+	Integer lat,
+	Integer lng,
 	Integer ang,
 	Integer spd
 ) {
